refacto: move into subfolder versions, take out constants, and merge conversions logic in create_case_health
