+++ converter/converter/versions/base_message_converter.py
@@ -10,7 +10,7 @@ class BaseMessageConverter:
             source_version_index = version_order_list.index(source_version)
         except ValueError:
             raise ValueError(f"Unknown source version {source_version}. Must be one of: {version_order_list}")
-    
+
         target_version_index = -1
         try:
             target_version_index = version_order_list.index(target_version)
@@ -20,7 +20,7 @@ class BaseMessageConverter:
         # No conversion needed
         if source_version_index == target_version_index:
             return edxl_json
-        
+
         try:
             # Convert directly if version are consecutive
             if abs(source_version_index - target_version_index) == 1:
@@ -40,13 +40,13 @@ class BaseMessageConverter:
             # Recursively call convert with the new source_version
             return self.convert(
                 version_order_list[source_version_index + version_delta],
-                target_version, 
+                target_version,
                 converted_edxl_json
             )
         except ValueError as err:
             print(f"[ERROR] {err}")
             self.raise_conversion_not_implemented_error(source_version, target_version)
-            
+
 
     def upgrade(self, source_version, source_version_index, edxl_json):
         if source_version == "v1":
@@ -55,7 +55,7 @@ class BaseMessageConverter:
             return self.convert_v2_to_v3(edxl_json)
         else:
             self.raise_conversion_impossible_error(source_version, version_order_list[source_version_index + 1])
-    
+
     def downgrade(self, source_version, source_version_index, edxl_json):
         if source_version == "v2":
             return self.convert_v2_to_v1(edxl_json)
@@ -66,13 +66,13 @@ class BaseMessageConverter:
 
     def convert_v1_to_v2(self, edxl_json):
         self.raise_conversion_not_implemented_error("v1", "v2")
-    
+
     def convert_v2_to_v1(self, edxl_json):
         self.raise_conversion_not_implemented_error("v2", "v1")
 
     def convert_v2_to_v3(self, edxl_json):
         self.raise_conversion_not_implemented_error("v2", "v3")
-    
+
     def convert_v3_to_v2(self, edxl_json):
         self.raise_conversion_not_implemented_error("v3", "v2")
