+++ converter/tests/test_v2_v3_converter.py
@@ -1,7 +1,7 @@
 import json
 from unittest.mock import patch
 from converter.constants import Constants
-from converter.v2_v3.create_case_health_converter import CreateHealthCaseConverter
+from converter.versions.create_case_health_converter import CreateHealthCaseConverter
 from .test_helpers import TestHelper, get_file_endpoint
 from snapshottest import TestCase
